==> scripts/06_animate_plays.Rmd <==
---
output:
  xaringan::moon_reader:
    self_contained: true
    nature:
      ratio: '16:9'
      highlightStyle: github
      highlightLines: true
      countIncrementalSlides: false
      slideNumberFormat: "%current%"
---

```{r setup, include=FALSE}
options(htmltools.dir.version = FALSE)
knitr::opts_chunk$set(
  # fig.width = 9,
  # fig.height = 3.5,
  fig.retina = 3, 
  out.width = "100%",
  # cache = TRUE,
  echo = FALSE,
  message = FALSE, 
  warning = FALSE,
  comment = "",
  hiline = TRUE,
  fig.align = "center"
)
library(tidyverse)
library(gganimate)

tracking <- arrow::read_parquet("../data/tracking.parquet")
plays <- read_csv("../data/plays.csv")

# source("scripts/04_fit_model.R")
# plays_snap_timing |> 
#   filter(passer_player_name == "P.Mahomes") |> 
#   left_join(select(players, nflId, displayName)) |> 
#   select(gameId, playId, cluster, 
#          motion_player = displayName, receiver_player_name, frame_between) |> 
#   inner_join(select(plays_use_motion_tag, gameId, playId))
```

---

## Example play animation

```{r}
# ex <- tracking |> 
#   filter(gameId == 2022092502, playId == 1180, nflId == 40011)
# ex |> 
#   select(frameId, x:event) |> 
#   drop_na() |> 
#   xtable::xtable()
```

.pull-left[

```{r}
ex1 <- tracking |> 
  filter(gameId == 2022092502, playId == 1180) |> 
  mutate(first_frame = frameId[which(event == "huddle_break_offense")][1],
         # first_frame = frameId[which(event == "line_set")][1],
         last_frame = frameId[which(event == "tackle")][1]) |> 
  filter(frameId >= first_frame & frameId <= last_frame) |> 
  mutate(
    pt_size = ifelse(club == "football", 2.7, 5.4),
    pt_fill_order = case_when(
      club == "IND" ~ "2",
      club == "KC" & nflId == "40011" ~ "3",
      club == "KC" & nflId != "40011" ~ "1",
      club == "football" ~ "4"
    )
  )
club1 <- ex1 |> 
  filter(club != "football")
football1 <- ex1 |>
  filter(club == "football")
path1 <- ex1 |>
  filter(nflId == 40011) |>
  mutate(motion_frame = frameId[which(event == "man_in_motion")][1],
         snap_frame = frameId[which(event == "ball_snap")][1]) |> 
  mutate(motion_color = "lightgray",
         motion_linetype = "dashed",
         motion_linewidth = 0.4)
motion1 <- path1 |>
  filter(frameId >= motion_frame & frameId <= snap_frame) |> 
  mutate(motion_color = "gold",
         motion_linetype = "solid",
         motion_linewidth = 1.2)
desc1 <- ex1 |> 
  distinct(gameId, playId) |> 
  left_join(plays) |> 
  mutate(playDescription = str_c("Q", quarter, ": ", playDescription),
         playDescription = str_remove_all(playDescription, "\\s\\([A-z]\\..*"),
         playDescription = str_replace_all(playDescription, "\\.", "\\. ")) |> 
  pull(playDescription)
```


```{r}
ggplot() +
  annotate("rect",
           xmin = 160/3,
           xmax = 0,
           ymin = 15,
           ymax = 65,
           fill = scales::alpha("#21ae5f", 0.9)) +
  annotate("text", 
           y = seq(20, 60, 10),
           x = 10,
           color = "white",
           family = "Chivo",
           label = seq(10, 50, 10),
           size = 6,
           angle = 90) +
  annotate("text", 
           y = seq(20, 60, 10),
           x = 40,
           color = "white",
           family = "Chivo",
           label = seq(10, 50, 10),
           size = 6,
           angle = 270) +
  annotate("text", 
           y = setdiff(seq(15, 65, 1), seq(15, 65, 5)),
           x = 160/3,
           color = "white",
           label = "—") +
  annotate("text", 
           y = setdiff(seq(15, 65, 1), seq(15, 65, 5)),
           x = 0,
           color = "white",
           label = "—") +
  annotate("text", 
           y = setdiff(seq(15, 65, 1), seq(15, 65, 5)),
           x = 23.36667,
           color = "white",
           size = 3,
           label = "—") +
  annotate("text", 
           y = setdiff(seq(15, 65, 1), seq(15, 65, 5)),
           x = 29.96667,
           color = "white",
           size = 3,
           label = "—") +
  annotate("segment", 
           y = 15,
           yend = 65,
           x = c(160/3, 0),
           xend = c(160/3, 0),
           color = "white") +
  geom_hline(yintercept = seq(15, 65, 5), color = "white") +
  annotate("segment", 
           y = 35,
           yend = 35,
           x = 0,
           xend = 160/3,
           size = 1.5,
           color = "#FDE725") +
  annotate("segment", 
           y = 27,
           yend = 27,
           x = 0,
           xend = 160/3,
           size = 1.5,
           #000aa0
           color = "midnightblue") +
  geom_point(data = club1, 
             aes(y, x, fill = pt_fill_order, size = pt_size, group = nflId), shape = 21) +
  geom_point(data = football1, 
             aes(y, x, size = pt_size, group = nflId), fill = "#654321", shape = 21) +
  geom_line(data = motion1,
            aes(y, x, color = motion_color, linetype = motion_linetype, linewidth = motion_linewidth)) +
  geom_line(data = path1, 
            aes(y, x, color = motion_color, linetype = motion_linetype, linewidth = motion_linewidth)) +
  scale_fill_manual(values = c("white", "#002C5F", "gold"),
                    labels = c("KC", "IND", "Motion (T. Kelce)")) +
  scale_size_identity() +
  scale_color_identity() +
  scale_linetype_identity() +
  scale_linewidth_identity() +
  scale_x_reverse() +
  transition_time(frameId) +
  transition_reveal(frameId) +
  ease_aes("linear") +
  guides(fill = guide_legend(override.aes = list(size = 5.4))) +
  labs(title = "Kansas City Chiefs @ Indianapolis Colts, 2022 NFL Week 3",
       subtitle = desc1,
       fill = NULL) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        plot.subtitle = element_text(size = 9.5, face = "italic", hjust = 0.5),
        text = element_text(family = "Fira Sans", color = "#26282A"),
        panel.grid = element_blank(),
        legend.position = "top",
        legend.text = element_text(size = 10.5),
        legend.margin = margin(b = -22),
        axis.text = element_blank(),
        axis.title = element_blank(),
        axis.ticks = element_blank())
```

]

.pull-right[

```{r}
ex2 <- tracking |> 
  filter(gameId == 2022102310, playId == 3897) |> 
  mutate(first_frame = frameId[which(event == "line_set")][1],
         last_frame = frameId[which(event == "touchdown")][1]) |> 
  filter(frameId >= first_frame & frameId <= last_frame) |> 
  mutate(
    pt_size = ifelse(club == "football", 2.7, 5.4),
    pt_fill_order = case_when(
      club == "SF" ~ "2",
      club == "KC" & nflId == "47839" ~ "3",
      club == "KC" & nflId != "47839" ~ "1",
      club == "football" ~ "4"
    )
  )

club2 <- ex2 |> 
  filter(club != "football")
football2 <- ex2 |>
  filter(club == "football")
motion2 <- ex2 |>
  filter(nflId == 47839) |>
  mutate(motion_frame = frameId[which(event == "man_in_motion")][1],
         snap_frame = frameId[which(event == "ball_snap")][1]) |>
  mutate(motion_color = ifelse(frameId >= motion_frame - 10 & frameId <= snap_frame, "gold", "lightgray"),
         motion_linetype = ifelse(frameId >= motion_frame - 10 & frameId <= snap_frame, "solid", "dashed"),
         motion_linewidth = ifelse(frameId >= motion_frame - 10 & frameId <= snap_frame, 1.2, 0.4))
desc2 <- ex2 |> 
  distinct(gameId, playId) |> 
  left_join(plays) |> 
  mutate(playDescription = str_c("Q", quarter, ": ", playDescription),
         playDescription = str_remove_all(playDescription, "\\.\\s.*"),
         playDescription = str_replace_all(playDescription, "\\.", "\\. ")) |> 
  pull(playDescription)
```


```{r}
ggplot() +
  annotate("rect",
           xmin = 160/3,
           xmax = 0,
           ymin = 53,
           ymax = 110,
           fill = scales::alpha("#21ae5f", 0.9)) +
  annotate("text", 
           y = seq(60, 110, 10),
           x = 10,
           color = "white",
           family = "Chivo",
           label = c(seq(50, 10, -10), "G   "),
           size = 6,
           angle = 90) +
  annotate("text", 
           y = seq(60, 110, 10),
           x = 40,
           color = "white",
           family = "Chivo",
           label = c(seq(50, 10, -10), "   G"),
           size = 6,
           angle = 270) +
  annotate("text", 
           y = c(53:54, setdiff(seq(55, 110, 1), seq(55, 110, 5))),
           x = 160/3,
           color = "white",
           label = "—") +
  annotate("text", 
           y = c(53:54, setdiff(seq(55, 110, 1), seq(55, 110, 5))),
           x = 0,
           color = "white",
           label = "—") +
  annotate("text", 
           y = c(53:54, setdiff(seq(55, 110, 1), seq(55, 110, 5))),
           x = 23.36667,
           color = "white",
           size = 3,
           label = "—") +
  annotate("text", 
           y = c(53:54, setdiff(seq(55, 110, 1), seq(55, 110, 5))),
           x = 29.96667,
           color = "white",
           size = 3,
           label = "—") +
  annotate("segment",
           y = 55,
           yend = 110,
           x = c(160/3, 0),
           xend = c(160/3, 0),
           color = "white") +
  geom_hline(yintercept = seq(55, 110, 5), color = "white") +
  annotate("segment",
           y = 70.4,
           yend = 70.4,
           x = 0,
           xend = 160/3,
           size = 1.5,
           color = "#FDE725") +
  annotate("segment",
           y = 64.7,
           yend = 64.7,
           x = 0,
           xend = 160/3,
           size = 1.5,
           color = "midnightblue") +
  geom_point(data = club2, 
             aes(y, x, fill = pt_fill_order, size = pt_size, group = nflId), shape = 21) +
  geom_point(data = football2, 
             aes(y, x, size = pt_size, group = nflId), fill = "#654321", shape = 21) +
  geom_line(data = motion2, 
            aes(y, x, color = motion_color, linetype = motion_linetype, linewidth = motion_linewidth)) +
  scale_fill_manual(values = c("white", "#AA0000", "gold"),
                    labels = c("KC", "SF", "Motion (M. Hardman)")) +
  scale_size_identity() +
  scale_color_identity() +
  scale_linetype_identity() +
  scale_linewidth_identity() +
  scale_x_reverse() +
  transition_time(frameId) +
  transition_reveal(frameId) +
  ease_aes("linear") +
  guides(fill = guide_legend(override.aes = list(size = 5.4))) +
  labs(title = "Kansas City Chiefs @ San Francisco 49ers, 2022 NFL Week 7",
       subtitle = desc2,
       fill = NULL) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        plot.subtitle = element_text(size = 9.5, face = "italic", hjust = 0.5),
        text = element_text(family = "Fira Sans", color = "#26282A"),
        panel.grid = element_blank(),
        legend.position = "top",
        legend.text = element_text(size = 10.5),
        legend.margin = margin(b = -22),
        axis.text = element_blank(),
        axis.title = element_blank(),
        axis.ticks = element_blank())
```


]

---

## Clustering features

```{r}
center1 <- ex1 |> 
  filter(nflId == 53492 & event == "man_in_motion") |> 
  mutate(pt_size = 8, x = x + 0.2)
event1 <- path1 |> 
  mutate(x_los = 27, 
         event = ifelse(frameId == frameId[which(x > x_los)][1], "cross_los", event)) |> 
  filter(event %in% c("man_in_motion", "ball_snap", "cross_los")) |> 
  mutate(pt_size = 8) |> 
  mutate(txt = c("Start of Motion", "Ball snap", "Crossing LOS"))


library(latex2exp)
ggplot() +
  annotate("rect",
           xmin = 29,
           xmax = 5,
           ymin = 19,
           ymax = 31,
           fill = scales::alpha("#21ae5f", 0.9)) +
  annotate("text", 
           y = seq(20, 30, 10),
           x = 10,
           color = "white",
           # family = "Chivo",
           label = seq(10, 20, 10),
           size = 6,
           angle = 90) +
  annotate("text", 
           y = c(setdiff(seq(20, 30, 1), seq(20, 30, 5))),
           x = 23.36667,
           color = "white",
           size = 3,
           label = "—") +
  geom_hline(yintercept = seq(20, 30, 5), color = "white") +
  annotate("segment", 
           y = 27,
           yend = 27,
           x = 5,
           xend = 29,
           linewidth = 1.5,
           #000aa0
           color = "midnightblue") +
  geom_path(data = filter(path1, frameId < 134), 
            aes(y, x, color = motion_color, linetype = motion_linetype), linewidth = 1) +
  geom_point(data = event1, 
             aes(y, x, fill = pt_fill_order, size = pt_size, group = nflId), shape = 21) +
  geom_point(data = center1, 
             aes(y, x, fill = pt_fill_order, size = pt_size, group = nflId), shape = 21) +
  geom_point(data = mutate(filter(football1, event == "man_in_motion"), pt_size = 3), 
             aes(y, x, size = pt_size, group = nflId), fill = "#654321", shape = 21) +
  ggrepel::geom_text_repel(data = event1, aes(y, x, label = txt), 
                           nudge_x = c(0, 1.5, 2.9), nudge_y = c(1, 0.9, 0.4), segment.color = NA,
                           family = "Fira Sans", fontface = "bold") +
  annotate("segment", 
           y = 23.2,
           yend = 23.2,
           x = 8,
           xend = 23.7,
           linewidth = 0.6,
           arrow = arrow(type = "closed", ends = "both", length = unit(0.015, "npc")),
           color = "black") +
  annotate("segment", 
           y = 24.5,
           yend = 24.5,
           x = 19.35,
           xend = 23.7,
           linewidth = 0.4,
           arrow = arrow(type = "closed", ends = "both", length = unit(0.015, "npc")),
           color = "black") +
  annotate("segment", 
           y = 28,
           yend = 28,
           x = 19.2,
           xend = 23.7,
           linewidth = 0.4,
           arrow = arrow(type = "closed", ends = "both", length = unit(0.015, "npc")),
           color = "black") +
  annotate("segment", 
           y = 26.3,
           yend = 24.5,
           x = 24.8,
           xend = 24.8,
           linewidth = 0.4,
           arrow = arrow(type = "closed", ends = "both", length = unit(0.015, "npc")),
           color = "black") +
  annotate("text", x = 15.5, y = 22.7, color = "white", family = "Fira Sans", size = 4.5,
           label = TeX("$\\textbf{M_{sideline}}$")) +
  annotate("text", x = 21.3, y = 24, color = "white", family = "Fira Sans", size = 4.5,
           label = TeX("$\\textbf{B_{sideline}}$")) +
  annotate("text", x = 26.5, y = 25.5, color = "white", family = "Fira Sans", size = 4.5,
           label = TeX("$\\textbf{B_{endzone}}$")) +
  annotate("text", x = 21.3, y = 28.5, color = "white", family = "Fira Sans", size = 4.5,
           label = TeX("$\\textbf{L_{sideline}}$")) +
  scale_fill_manual(values = c("white", "gold"),
                    labels = c("Center (C. Humphrey)", "Motion (T. Kelce)")) +
  scale_size_identity() +
  scale_color_identity() +
  scale_linetype_identity() +
  scale_linewidth_identity() +
  scale_x_reverse() +
  guides(fill = guide_legend(override.aes = list(size = 5.4))) +
  labs(fill = NULL) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        plot.subtitle = element_text(size = 9.5, face = "italic", hjust = 0.5),
        plot.background = element_rect(fill = "white", color = NA),
        text = element_text(family = "Fira Sans", color = "#26282A"),
        panel.grid = element_blank(),
        legend.position = "top",
        legend.text = element_text(size = 10.5),
        legend.margin = margin(b = -18),
        axis.text = element_blank(),
        axis.title = element_blank(),
        axis.ticks = element_blank())

# ggsave(
#   "figures/clustering_features.png",
#   width = 5.6,
#   height = 3.6,
#   dpi = 1200
# )
```


---

## Create snapshots

```{r}
plot_field <- function(x) {
  
  ggplot() +
    annotate("rect",
             xmin = 160/3,
             xmax = 0,
             ymin = 15,
             ymax = 45,
             fill = scales::alpha("#21ae5f", 0.8)) +
    annotate("text", 
             y = seq(20, 40, 10),
             x = 10,
             color = "white",
             # family = "Chivo",
             label = seq(10, 30, 10),
             size = 3.5,
             angle = 90) +
    annotate("text", 
             y = seq(20, 40, 10),
             x = 40,
             color = "white",
             # family = "Chivo",
             label = seq(10, 30, 10),
             size = 3.5,
             angle = 270) +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 160/3,
             color = "white",
             alpha = 0.6,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 0,
             color = "white",
             alpha = 0.6,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 23.36667,
             color = "white",
             size = 2.5,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 29.96667,
             color = "white",
             size = 2.5,
             label = "—") +
    annotate("segment", 
             y = 15,
             yend = 45,
             x = c(160/3, 0),
             xend = c(160/3, 0),
             color = "white") +
    annotate("segment", 
             y = seq(15, 45, 5),
             yend = seq(15, 45, 5),
             x = 160/3,
             xend = 0,
             linewidth = 0.3,
             color = "white") +
    # geom_hline(yintercept = seq(15, 45, 5), color = "white") +
    annotate("segment", 
             y = 35,
             yend = 35,
             x = 0,
             xend = 160/3,
             size = 1.5,
             color = "#FDE725") +
    annotate("segment", 
             y = 27,
             yend = 27,
             x = 0,
             xend = 160/3,
             size = 1.5,
             #000aa0
             color = "midnightblue") +
    geom_path(data = filter(path1, frameId >= 59), 
              aes(y, x, color = motion_color, linetype = motion_linetype), linewidth = 0.5) +
    geom_line(data = motion1,
              aes(y, x, color = motion_color, linetype = motion_linetype), linewidth = 0.5) +
    # geom_line(data = filter(path1, event == {{x}}), 
    #           aes(y, x, color = motion_color, linetype = motion_linetype, linewidth = motion_linewidth)) +
    geom_point(data = filter(club1, club == "IND" & event == {{x}}), 
               aes(y, x, fill = pt_fill_order, size = pt_size*0.54, group = nflId), shape = 21) +
    geom_point(data = filter(club1, club == "KC" & event == {{x}}), 
               aes(y, x, fill = pt_fill_order, size = pt_size*0.6, group = nflId), shape = 21) +
    geom_point(data = filter(football1, event == {{x}}), 
               aes(y, x, size = pt_size*0.65, group = nflId), fill = "#654321", shape = 21) +
    scale_fill_manual(values = c("white", "#002C5F", "gold"),
                      labels = c("KC", "IND", "Motion (T. Kelce)")) +
    scale_size_identity() +
    scale_color_identity() +
    scale_linetype_identity() +
    scale_linewidth_identity() +
    scale_x_reverse() +
    labs(title = {{x}}) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = 0.5, face = "bold", size = 11, margin=margin(0,0,0,0)),
          plot.subtitle = element_text(size = 9.5, face = "italic", hjust = 0.5),
          text = element_text(family = "Fira Sans", color = "#26282A"),
          plot.margin=grid::unit(c(3,-3,-1,-3), "mm"),
          panel.grid = element_blank(),
          legend.position = "none",
          axis.text = element_blank(),
          axis.title = element_blank(),
          axis.ticks = element_blank())
}
```

```{r}
plot_field_first <- function(x) {
  
  ggplot() +
    annotate("rect",
             xmin = 160/3,
             xmax = 0,
             ymin = 15,
             ymax = 45,
             fill = scales::alpha("#21ae5f", 0.8)) +
    annotate("text", 
             y = seq(20, 40, 10),
             x = 10,
             color = "white",
             # family = "Chivo",
             label = seq(10, 30, 10),
             size = 3.5,
             angle = 90) +
    annotate("text", 
             y = seq(20, 40, 10),
             x = 40,
             color = "white",
             # family = "Chivo",
             label = seq(10, 30, 10),
             size = 3.5,
             angle = 270) +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 160/3,
             color = "white",
             alpha = 0.6,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 0,
             color = "white",
             alpha = 0.6,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 23.36667,
             color = "white",
             size = 2.5,
             label = "—") +
    annotate("text", 
             y = setdiff(seq(15, 45, 1), seq(15, 45, 5)),
             x = 29.96667,
             color = "white",
             size = 2.5,
             label = "—") +
    annotate("segment", 
             y = 15,
             yend = 45,
             x = c(160/3, 0),
             xend = c(160/3, 0),
             color = "white") +
    annotate("segment", 
             y = seq(15, 45, 5),
             yend = seq(15, 45, 5),
             x = 160/3,
             xend = 0,
             linewidth = 0.3,
             color = "white") +
    # geom_hline(yintercept = seq(15, 45, 5), color = "white") +
    annotate("segment", 
             y = 35,
             yend = 35,
             x = 0,
             xend = 160/3,
             size = 1.5,
             color = "#FDE725") +
    annotate("segment", 
             y = 27,
             yend = 27,
             x = 0,
             xend = 160/3,
             size = 1.5,
             #000aa0
             color = "midnightblue") +
    geom_point(data = filter(club1, club == "IND" & event == {{x}}), 
               aes(y, x, fill = pt_fill_order, size = pt_size*0.54, group = nflId), shape = 21) +
    geom_point(data = filter(club1, club == "KC" & event == {{x}}), 
               aes(y, x, fill = pt_fill_order, size = pt_size*0.6, group = nflId), shape = 21) +
    geom_point(data = filter(football1, event == {{x}}), 
               aes(y, x, size = pt_size*0.65, group = nflId), fill = "#654321", shape = 21) +
    scale_fill_manual(values = c("white", "#002C5F", "gold"),
                      labels = c("KC", "IND", "Motion (T. Kelce)")) +
    scale_size_identity() +
    scale_color_identity() +
    scale_linetype_identity() +
    scale_linewidth_identity() +
    scale_x_reverse() +
    labs(title = {{x}}) +
    theme_minimal() +
    theme(plot.title = element_text(hjust = 0.5, face = "bold", size = 11, margin=margin(0,0,0,0)),
          plot.subtitle = element_text(size = 9.5, face = "italic", hjust = 0.5),
          text = element_text(family = "Fira Sans", color = "#26282A"),
          plot.margin=grid::unit(c(3,-3,-1,-3), "mm"),
          panel.grid = element_blank(),
          legend.position = "none",
          axis.text = element_blank(),
          axis.title = element_blank(),
          axis.ticks = element_blank())
}
```

```{r}
library(cowplot)
fig_field <- plot_grid(
  plot_field_first("huddle_break_offense"),
  plot_field_first("line_set"),
  plot_field("man_in_motion"),
  plot_field("ball_snap"),
  plot_field("pass_outcome_caught"),
  plot_field("tackle"),
  nrow = 3
)

ggsave(
  "figures/field_snapshots.png",
  plot = fig_field,
  width = 7.8,
  height = 8.4,
  dpi = 1200,
  bg = "white"
)
```
